refactor(auth): split openai.py into an auth/openai package
Mechanical, behavior-preserving split by responsibility: constants,
catalog, oauth_client, browser_flow, models, config_apply, login, plus
a package __init__ re-exporting the existing import surface. Applies
the accepted clean-code review items: JsonObject boundary typing with
narrow casts, _parse_chatgpt_model_item extraction,
_default_config_error / _handled_error_event helpers, and
private-helper renames (_first_present_field /
_first_non_empty_string_field). Test monkeypatch targets follow the
moved definitions. Includes the audit's 401/403 login message split,
which now lives in login.py.
